Add a way to add/remove block

+++ Source/MineUE4/BlockArray.cpp
@@ -6,8 +6,8 @@
 
 bool FCompressedBlock::NetSerialize(FArchive& Ar, class UPackageMap* Map, bool& bOutSuccess)
 {
-  static const uint16 NMBBITSMAXTYPEOFBLOCKS = (uint16)FMath::Log2(AChunk::MAXTYPEOFBLOCKS - 1) + 1;
-  static const uint32 NMBMAXBBLOCKS = (uint16)FMath::Log2(AChunk::CHUNKSIZEX * AChunk::CHUNKSIZEY * AChunk::CHUNKSIZEZ - 1) + 1;
+  static const int32 NMBBITSMAXTYPEOFBLOCKS = (int32)FMath::Log2(AChunk::MAXTYPEOFBLOCKS - 1) + 1;
+  static const int32 NMBMAXBBLOCKS = (int32)FMath::Log2(AChunk::CHUNKSIZEX * AChunk::CHUNKSIZEY * AChunk::CHUNKSIZEZ - 1) + 1;
 
   //Max is 255 inclusive, so we need to only serialize only 8 bits
   Ar.SerializeBits(&BlockType, NMBBITSMAXTYPEOFBLOCKS);
